refactor: move note creation logic from AppSidebar to NotebookView
- Move handleNewNote from AppSidebar to handleAddNoteWindow in NotebookView
- Remove direct note creation logic from AppSidebar component
- Add onAddNote prop to AppSidebar for proper separation of concerns
- Remove notebookId prop from AppSidebar (no longer needed)
- Clean up unused NoteEditorPayload import in AppSidebar
- Rename handleAddWindow to handleAddBrowserWindow for consistency

This follows the same pattern as chat/browser window creation where the
parent component handles entity creation and the sidebar only triggers actions.
